# spec/api.yaml
### Generated-by: Spec-kit
### Generator: spec-kit
### Do not edit generated sections without updating `spec/api.yaml` source
openapi: 3.0.3
info:
  title: Petstore API
  version: '1.0.0'
  description: OpenAPI spec generated by Spec-kit agent for the Petstore service.
servers:
  - url: http://localhost:8080
paths:
  /api/pets:
    get:
      summary: List pets
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: species
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
    post:
      summary: Create pet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pet'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
  /api/pets/{id}:
    get:
      summary: Get pet by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '404':
          description: Not found
    put:
      summary: Update pet
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pet'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '404':
          description: Not found
    delete:
      summary: Delete pet
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
components:
  schemas:
    Pet:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        species:
          type: string
        age:
          type: integer
        description:
          type: string
        imageUrl:
          type: string
        price:
          type: number
          format: double
      required:
        - name
        - species